+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # Unreleased
 
+* Fix kubernetes discovery module to not override user defined tasks.
+
 # 0.25.1
 
 * Improve GitHub API client management. Won't re-auth constantly if cache is disabled, and will keep the connection pool on re-auth.